+++ commands/join.js
@@ -106,7 +106,7 @@ exports.run = (client, message, args) => {
                     let filledMessage = new FutureMessage(pingMessage, createEmbed(client,"Squad filled",`Squad ${squads[i]} has been filled\nOriginal message: ${currentSquad.messageContent}`))
                     futureMessages.push(filledMessage);
 
-                    fillSquad(client, currentSquad.lobbyID);
+                    fillSquad(client, currentSquad.lobbyID, message.channel);
                 }
             }
             
@@ -170,31 +170,74 @@ exports.run = (client, message, args) => {
         recruitChatChannel.send(newMessage.message, newMessage.embed);
     }
     
-    doEdits(client, editMessages, message);
+    doEdits(client, editMessages, message.channel);
+
+    message.delete(5000)
+    .catch(() => {
+        let catchMessage = 'Handled rejection - caught in Join - success'
+        console.log(catchMessage);
+
+        let logChannel = client.channels.find(channel => channel.id === client.config.get('channelConfig').logChannel);
+        logChannel.send(`<@198269661320577024>, ${catchMessage}`);
+    });
     
 };
 
-function fillSquad(client, id) {
+async function fillSquad(client, id, channel) {
     closeSquad(client, id);
 
-    closeOthers(client, id);
+    let thisSquad = client.lobbyDB.get(id);
 
-    //pullPlayers(client, id);
+    let squadPlayers = [];
+    squadPlayers.push(thisSquad.hostID);
+    for (player of thisSquad.joinedIDs) squadPlayers.push(player);
 
+    //for each player
+    for (player of squadPlayers) {
+        //close all
+        closeOthers(client, player);
+    }
+
+    for (player of squadPlayers) {
+        //leave all
+        pullPlayers(client, player, channel);
+    }
 }
 
-function closeOthers(client, id) {
-    //get current host
-    let oldSquad = client.lobbyDB.get(id);
-    let thisHost = oldSquad.hostID;
+function pullPlayers(client, player, channel) {
+    //editMessages.push({messageID: squad.messageID, messageIndex: squad.countIndex, count: squad.playerCount, lobbyID: squad.lobbyID});
+    let editMessages = [];
 
+    //find all other squads they're in
+    for (let i = 0; i < client.config.get('baseConfig').maxSquads; i++) {
+        //(if the squad ID exists)
+        if (client.lobbyDB.has(i.toString())) {
+            let squad = client.lobbyDB.get(i.toString());
+            if (!squad.open) continue;
+            //if they're in the squad
+            if (squad.joinedIDs.includes(player)) {
+                //leave it
+                squad.playerCount--;
+                squad.joinedIDs.splice(squad.joinedIDs.indexOf(player), 1);
+
+                client.lobbyDB.set(i.toString(), squad);
+
+                editMessages.push({messageID: squad.messageID, messageIndex: squad.countIndex, count: squad.playerCount, lobbyID: squad.lobbyID});
+            }
+        }
+    }
+
+    doEdits(client, editMessages, channel);
+}
+
+function closeOthers(client, playerID) {
     //find all other squads they're hosting
     for (let i = 0; i < client.config.get('baseConfig').maxSquads; i++) {
         //(if the squad ID exists)
         if (client.lobbyDB.has(i.toString())) {
             let squad = client.lobbyDB.get(i.toString());
             //if they're the same host
-            if (squad.hostID == thisHost) {
+            if (squad.hostID == playerID) {
                 //close it
                 closeSquad(client, i.toString());
             }
@@ -203,11 +246,17 @@ function closeOthers(client, id) {
 }
 
 async function closeSquad (client, id) {
-    //close the squad
-    client.lobbyDB.setProp(id, "open", false);
 
     //get the squad
     const squad = client.lobbyDB.get(id);
+
+    //check if already closed
+    if (!squad.open) return;
+
+    //close the squad
+    client.lobbyDB.setProp(id, "open", false);
+
+    
     //delete the host message
     //SPLIT RECRUITS
     //const channel = squad.channel;
@@ -215,13 +264,19 @@ async function closeSquad (client, id) {
     const messageID = squad.messageID;
 
     const channel = client.channels.get(channelID);
-    channel.fetchMessage(messageID)
+    let messageNotFound = false;
+    await channel.fetchMessage(messageID)
+    .catch(() => {
+        messageNotFound = true;
+        let logChannel = client.channels.find(channel => channel.id === client.config.get('channelConfig').logChannel);
+        logChannel.send(`<@198269661320577024> Error deleting message for squad ${id} for message ID ${messageID}. Does it exist?`);
+    })
     .then(squadMessage => {
-        squadMessage.delete();
+        if (!messageNotFound) squadMessage.delete();
     })
 }
 
-async function doEdits(client, editMessages, message) {
+async function doEdits(client, editMessages, channel) {
     const { Client, RichEmbed } = require('discord.js');
 
     let currentMessage = null;
@@ -233,7 +288,7 @@ async function doEdits(client, editMessages, message) {
 
         if (currentMessage == null || currentMessage.id != edit.messageID) {
 
-            currentMessage = await message.channel.fetchMessage(edit.messageID)
+            currentMessage = await channel.fetchMessage(edit.messageID)
             .catch(() => {
                 messageNotFound = true;
                 let logChannel = client.channels.find(channel => channel.id === client.config.get('channelConfig').logChannel);
@@ -257,14 +312,7 @@ async function doEdits(client, editMessages, message) {
         await currentMessage.edit(embed);
     }
 
-    message.delete(5000)
-    .catch(() => {
-        let catchMessage = 'Handled rejection - caught in Join - success'
-        console.log(catchMessage);
-
-        let logChannel = client.channels.find(channel => channel.id === client.config.get('channelConfig').logChannel);
-        logChannel.send(`<@198269661320577024>, ${catchMessage}`);
-    });
+    
 }
 
 function createEmbed(client, title, content) {
